+++ packages/types/src/catalog.ts
@@ -4,8 +4,12 @@ export type Catalog =
   Flowglad.CustomerProfileRetrieveBillingResponse['catalog']
 
 export type Product =
-  Flowglad.CustomerProfileRetrieveBillingResponse.Catalog.Product
+  Flowglad.CustomerProfileRetrieveBillingResponse.Catalog.Product.Product
 
-export type Variant =
-  | Flowglad.CustomerProfileRetrieveBillingResponse.Catalog.Product.SinglePaymentVariant
-  | Flowglad.CustomerProfileRetrieveBillingResponse.Catalog.Product.SubscriptionVariant
+export type SinglePaymentVariant =
+  Flowglad.CustomerProfileRetrieveBillingResponse.Catalog.Product.SinglePaymentVariant
+
+export type SubscriptionVariant =
+  Flowglad.CustomerProfileRetrieveBillingResponse.Catalog.Product.SubscriptionVariant
+
+export type Variant = SinglePaymentVariant | SubscriptionVariant