+++ packages/types/src/invoice.ts
@@ -1,6 +1,13 @@
 import { type Flowglad } from '@flowglad/node'
 
+export type PurchaseInvoice =
+  Flowglad.Invoice.InvoiceRetrieveResponse.PurchaseInvoice
+
+export type SubscriptionInvoice =
+  Flowglad.Invoice.InvoiceRetrieveResponse.SubscriptionInvoice
+
 export type Invoice = Flowglad.Invoice.InvoiceRetrieveResponse
+
 export type InvoiceLineItem = Flowglad.InvoiceLineItemRetrieveResponse
 export type InvoiceStatus =
   Flowglad.Invoice.InvoiceRetrieveResponse['status']