+++ packages/types/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @flowglad/types
 
+## 0.4.21
+
+### Patch Changes
+
+- types: more exported discriminated union types
+
+## 0.4.20
+
+### Patch Changes
+
+- Product export fix
+
+## 0.4.19
+
+### Patch Changes
+
+- types: Currency -> CurrencyCode
+
 ## 0.4.18
 
 ### Patch Changes